Added auto-snapshot for -break and -step modes

+++ src/ANSTE/Virtualizer/Virt.pm
@@ -20,6 +20,7 @@ use base 'ANSTE::Virtualizer::Virtualizer';
 use strict;
 use warnings;
 
+use ANSTE;
 use ANSTE::Config;
 use ANSTE::Image::Image;
 use ANSTE::Exceptions::MissingArgument;
@@ -204,6 +205,7 @@ sub destroyImage
     defined $image or
         throw ANSTE::Exceptions::MissingArgument('image');
 
+    $self->deleteSnapshot($image, ANSTE::snapshotName());
     $self->execute("virsh destroy $image");
 }
 
@@ -672,6 +674,10 @@ sub createSnapshot
 {
     my ($self, $domain, $name, $description) = @_;
 
+    if ($self->existsSnapshot($domain, $name)) {
+        $self->deleteSnapshot($domain, $name);
+    }
+
     $self->execute("virsh snapshot-create-as $domain $name '$description'") or
         throw ANSTE::Exceptions::Error("Error creating snapshot $name in domain $domain");
 }
@@ -706,7 +712,9 @@ sub deleteSnapshot
 {
     my ($self, $domain, $name) = @_;
 
-     $self->execute("virsh snapshot-delete $domain $name") or
+    return unless $self->existsSnapshot($domain, $name);
+
+    $self->execute("virsh snapshot-delete $domain $name") or
         throw ANSTE::Exceptions::Error("Error deleting snapshot $name in domain $domain");
 }
